backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
pydantic
pydantic-settings